ir overview components
